Added Collada support, updated data model
PAKTool now exports as DAE as well as OBJ. Still no textures.
I discovered a couple more lists in some 3DO1 files, so I added them. Not sure what they are for, though it's likely one has something to do with subobjects.

I also discovered that one of the unknown ints in the main header points to a block that I hadn't noticed. It only appears in some files, after the 3do files and before the textures. No idea what that is.